Inform object_sets about problems in attributes
When something unexpected happens in one of the attributes or objects we
want to let users know about it.
We might be able to process affected attribute or object, but that
should decrease user's trust in data.
Hence we need to store and report errors coming from object sets also.