Let user choose exception level
We detect various specification violation. Some of them we consider
insignificant enough and recover.
Out users can, however, disagree with us about importance of certain
specification violations.
Hence we give users a little bit more control over it.

For example, users want to receive an error everytime something is
spoiled even a little bit.
Or on the opposite - users may want to get as much data as possible,
accepting the danger that data might be spoiled.