Kill custom user- and runtime_warning
With the introduction of dl::error_handler these warnings are surplus
and not used. The only exeception is the runtime_warning which is used
once in the binding code. This should be exchanged for dl::error_handler
too, but for for now just move the runtime_warning to core.cpp