Merge the extension- and dlisio-target into one
It's time to say goodbye to the extension-target. This commit kills the
extension by merging it into the dlisio target. Most of the noise from
this commit is from restructuring some of the header- and implementation
files, which was necessary to avoid naming conflicts.  Everything
related to the custom dlis types (i.e. snorm, unorm, etc) are moved into
types.hpp/types.cpp while the remaining parts of the old ext/types.hpp
and parse.cpp are moved into explicits.*