Add LIS documentation to sphinx
This commit adds documentation of the LIS reader to the sphinx
documentation.

The fixed records, implented in C++ with python bindings by pybind11,
are documented manually in docs/lis/api.rst. Sphinx pics up the docs
from the compiled pybind11 python module, and not directly from the
source files. Unfortunally it's not straight forward to build dlisio in
readthedocs' pipeline.

There are several ways around this, all with their own drawbacks. For now
we do the easy solution of manually documenting the classes from the
pybind11 module, but this is subject to change as we implement more part
of the LIS specification.